Realize that my algorithm is broken
As is made immediately obvious by the main app right now, I can't reduce
the identity function. My beta reduce function assumes that the
substitution happens in place for the root node, but this isn't true if
the root node is the bound variable.

More fundamentally, the problem is that I don't want to make copies of
everything. I only want to make copies of terms that involve the
substituted argument. In particular, existing variables should
definitely not be copied, since this might unbind a variable that is
bound somewhere above root!