Update debug prints, identify bug
Right now when a lambda inside the redex function body uses the outer
bound variable (i.e. the substituted one), the inner lambda's bound
variable gets reused by a new lambda term. This is bad.

Also the fact that I'm not detecting this situation and throwing an
error implies I'm doing something wrong in how I splice in the new node
that results from the redex function. In particular, if reduce_term
returns a LambdaTerm (as opposed to a TermId), its children don't know
about it (since it doesn't have an id to add to their parent lists). So
I need to fix this too.